FEAT-7477: Move the query compliler caster so it can be used in more places Move the query casting calculator out to the base directory so it can be used by the API layer as well.
