+++ modin/tests/pandas/native_df_interoperability/test_compiler_caster.py
@@ -14,12 +14,12 @@
 import pandas
 import pytest
 
+from modin.core.storage_formats.base.query_compiler_calculator import (
+    QueryCompilerCostCalculator,
+)
 import modin.pandas as pd
 from modin.core.storage_formats.base.query_compiler import QCCoercionCost
 from modin.core.storage_formats.pandas.native_query_compiler import NativeQueryCompiler
-from modin.core.storage_formats.pandas.query_compiler_caster import (
-    QueryCompilerCasterCalculator,
-)
 
 
 class CloudQC(NativeQueryCompiler):
@@ -229,7 +229,7 @@ def test_default_to_caller(default_df, default2_df):
 
 
 def test_no_qc_data_to_calculate():
-    calculator = QueryCompilerCasterCalculator()
+    calculator = QueryCompilerCostCalculator()
     calculator.add_query_compiler(ClusterQC)
     result = calculator.calculate()
     assert result is ClusterQC
@@ -237,7 +237,7 @@ def test_no_qc_data_to_calculate():
 
 
 def test_no_qc_to_calculate():
-    calculator = QueryCompilerCasterCalculator()
+    calculator = QueryCompilerCostCalculator()
     with pytest.raises(ValueError):
         calculator.calculate()
